news_api/requirements.txt
asgiref==3.8.1
attrs==23.2.0
cffi==1.16.0
cryptography==42.0.8
Django==4.1.13
django-filter==22.1
django-redis==5.4.0
djangorestframework==3.15.1
djongo==1.3.6
dnspython==2.6.1
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
packaging==24.1
pycparser==2.22
pymongo==4.7.3
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzstd==0.16.0
redis==5.0.6
referencing==0.35.1
rpds-py==0.18.1
setuptools==70.0.0
sqlparse==0.2.4
swagger-spec-validator==3.0.3
typing_extensions==4.12.2
uritemplate==4.1.1
whitenoise==6.7.0
